More clarity on types, implement type decoding
By analogy to Python itself, encoding means converting to a Python type
to a QPDFObjectHandle and decoding is back to Python.

A few native types can be transparently converted. What remains is
adding C++ wrappers for the types that have no native equivalent in
Python, the nonscalars, etc. Might need some metaclass magic or ability
to import Python helpers.