Implement PDF/A repair
qpdf omits the optional \n in front of endstream. It’s not required for
regular PDF but is required in PDF/A. Go through and add them
unconditionally by jumping into a helper Python module.

This was a pain, since it required adding a micro PDF parser of sorts.
Unfortunately there’s no way to hook qpdf in all the right places, so
this was the only way.

Should be disabled when not saving a PDF/A.