Move PageList implementation to C++
Not entirely sure why it didn't work in its current form, but decided it would be better to implement PageList in C++ since it would reduce possibility for reference counting errors at the boundaries, in conjunction with how QPDF::addPageAt uses a reference to a temporary page. Worked like a charm.