Nevermind, it's not fixed
It appears to only be fixed when the pyqpdf.inc_ref() is hacked in via custom type conversion being compiled at the time of the commit that claims it's fixed. (Build system doesn't notice changes in headers.)

This latest effort attempts to use reference_internal everywhere to force keep alives.