Use keep_alive_impl instead of inc_ref
We're not supposed to touch it, of course, but it may be the best option for now. An alternative is implementing a private version of the same thing.

keep_alive_impl preserves the owner PDF as long as related handles exist in Python. Probably still leaks pages copied from one PDF to another.