Improve type_caster
pybind11 people indicated that all three cast() overloads of type_caster_base really ought to be implemented. This borrows ideas from the PYBIND11_TYPE_CASTER macro.

We implement a rvalue caster that just defers to the pointer caster with a move instruction, and a lvalue caster that likes to copy. Removing the 'if' statement will cause a lot crashes.

As of this commit bindings that attempt to return QPDFObjectHandle* will fail to compile. If allowed, attempts to always crash. But there's little reason to allocate those handles on the heap so it actually makes sense in our usage to block any attempt to return them and ensure that we're copying.