Complete vendoring of pybind11
pybind11's development has slowed and essential fixes we need
are not in a public release. Vendor it so we can get more recent
versions.

Did this in such a way that nuking the src/vendor/pybind11 at some time
will let us drop vendoring in the future.

In particular we need:
https://github.com/pybind/pybind11/commit/b4719a60d32821fa47aaaba7c1035cb447bab386
(3.7 TSS API)

and
https://github.com/pybind/pybind11/commit/cbd16a8247b65c6eaab0369a54a25f51c5c0bd45
(stl.h type-specific caster fix)