Back to objectmaps again - a few fixes
Now thinking that ObjectMap is too drastic because there are
situations where we might not want an opaque object.

For example .images returns an ObjectMap with this turned on, but
that is really a view....

Several tests fail and assumptions about the nature of dictionary keys
are violated.

Note: Pulled some previously cherry-picked items of this branch when rebasing.