Improve const-correctness policy
Change convention to passing QPDFObjectHandle by value, removing
some cases where we use const_cast to add const and some
cases where a reference was used for a efficiency alone. It seems
more clear to pass by value as a
way of expressing that the passed object is not modified.

At the same time, enforce const more strongly for code that deals
with pybind11 which is const-correct.

Exempt from this change is our custom type converter.