Sort .cpp build inputs to ensure a reproducible build
Whilst working on the Reproducible Builds effort [0] we noticed that
pikepdf could not be built reproducibly.

This is due to the .cpp input files were compiled/linked in an order that was
determined by their layout on the filesystem which is, alas, non-deterministic.

This was originally filed in Debian as #950138 [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/950138