Limit Decimal object precision
When floats are converted to decimal the results can be excessively
high precision, to the point that it upsets Acrobat. Here we limit the
precision of Decimal objects we generate to 15 digits, which ought to
be enough. Since generating files with this much precision is generally
an error we treat it as a bug fix.

A semi-private allows manipulating the internal precision.