Properly register all C++ ObjectHelper subclasses
With what we did previously, pybind11 could not see that they were
subclasses of ObjectHelper so did not have proper type information.

QPDFFileSpecObjectHelper/AttachedFileSpec requires a shared_ptr
holder to avoid segfaults apparently due to it retain a reference to QPDF.
It appears that pybind11 requires all members of a class hierarchy to
use the same holder type, which is of course eminently reasonable. As far as I can
this is also not documented and the error message when this constraint
is violated, in our case, blames an related py::enum.

In any event, using shared_ptr for all of these even though only one
needs it, is not a properly. Hopefully?