Don't install psutil on Windows and update other test dependencies
For Windows + PyPy, we need to manually compile psutil, which fails in
some cases, but is a pointless exercise anyway since the relevant
test code does not work on Windows.

It's easier to just remove Windows as a test platform for psutil since POSIX
is sufficient to confirm correctness.

Also drop numpy limit on cp312 since they have released it.