Improve Object.__bool__ checks
In some cases, pikepdf.Object would throw an exception when perform a test such as

if pikepdf.Name('/Foo'):

with the reason being that no __bool__ was defined, so Python called __len__, and __len__ raised an exception.

In many of these cases there is no meaningful len(), because of PDF
encoding quirks. But we can define bool and make certain checks work better.