Reconfigure macOS build to use gnutls instead of openssl
Users reported trouble with open legacy encrypted files on macOS
specifically, e.g. https://github.com/pikepdf/pikepdf/issues/520

It appears this is because we were using Homebrew's qpdf which is
currently linked against Homebrew's openssl, which is phasing out
legacy crypto and needs to be compiled in a special mode to enable it.
Understandable but inconvenient for us.

So now we build and link against our own libqpdf, which in turn is
linked against gnutls, which does not seem to have the same issues.
All of our Linux and macOS builds are doing the same thing now,
rather than being split on crypto provider.